+++ src/utils/bech32-manager.js
