+++ src/utils/index.js
@@ -1,16 +1,16 @@
-import bech32Manager from './bech32Manager';
-import blocksRequestHelper from './blocksRequestHelper';
-import coinAdapter from './coinAdapter';
-import dateHandler from './dateHandler';
-import msgBuilder from './msgBuilder';
-import proposerHandler from './proposerHandler';
-import regExpBuilder from './regExpBuilder';
-import storageHandler from './storageHandler';
-import stringEncoder from './stringEncoder';
-import tokensHandler from './tokensHandler';
-import txHandler from './txs/txHandler';
-import txsFilter from './txs/txsFilter';
-import txsTableAdapter from './txs/txsTableAdapter';
+import bech32Manager from './bech32-manager';
+import blocksRequestHelper from './blocks-request-helper';
+import coinAdapter from './coin-adapter';
+import dateHandler from './date-handler';
+import msgBuilder from './msg-builder';
+import proposerHandler from './proposer-handler';
+import regExpBuilder from './reg-exp-builder';
+import storageHandler from './storage-handler';
+import stringEncoder from './string-encoder';
+import tokensHandler from './tokens-handler';
+import txHandler from './txs/tx-handler';
+import txsFilter from './txs/txs-filter';
+import txsTableAdapter from './txs/txs-table-adapter';
 
 export {
   bech32Manager,
