+++ src/utils/reg-exp-builder.js
