+++ src/utils/string-encoder.js
